requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-crispy-forms==1.14.0
django-environ==0.8.1
idna==3.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pymongo==4.1.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
